Closes peek/peek-performance_bar#16: Fixing incorrect time on initial page load with Turbolinks
In Turbolinks 2.0 the `page:change` event was changed to be fired on initial page load, which caused the time output to be incorrect because it wasn't actually a Turbolinks request. The solution appears to be to hook into the `page:load` event, which is only fired when a Turbolinks request is actually made.